gpui_macos: Drop redundant u64 casts in the GPU stats logging
allocated_size() and current_allocated_size() already return u64.

+++ crates/gpui_apple/src/metal_atlas.rs
@@ -38,7 +38,7 @@ impl MetalAtlas {
             list.textures
                 .iter()
                 .flatten()
-                .map(|texture| texture.metal_texture.allocated_size() as u64)
+                .map(|texture| texture.metal_texture.allocated_size())
                 .sum()
         }
         let lock = self.0.lock();
